Develop refactoring

+++ Assets/Scripts/Scenes/BattleScene/Interface/BossSkills/IBossRushAttack.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public interface IBossRushAttack 
 {
     public bool IsAttack { get; set; }
-    public void Attack(float basicSpeed, int basicDamage);
+    public void Attack(float basicSpeed, float basicDamage);
     public void StopAttack();
 
 }
