+++ Assets/Scripts/Scenes/BattleScene/Interface/BossSkills/IBossWorldAreaAttack.cs
@@ -5,6 +5,9 @@ using UnityEngine;
 public interface IBossWorldAreaAttack 
 {
     public bool IsAttack { get; set; }
-    public void Attack(int bossDamage, int animHash);
+
+    public void Attack(float bossDamage, int animHash);
+
     public void StopAttack();
+
 }
